Some changes to the Video classes.
Moved to a composition refactoring instead of abstract class. This
was to support use of attrs which will help with serialization I
think. No major changes to the API other than object construction
is now done via class methods. Maybe later on down the line Video
should really be changed to VideoReader or something since it is
read only for now. Improved the documentation and added some MP4
tests.