Modify quiverplot test to count points rather than child elements
We had been creating Qt element for each arrow, and the test checked
these. Now we're passing an array of points to the painter, which is
must faster. The test now checks the number of points in the array.
This isn't ideal, since we aren't making sure that the points actually
appear on the screen.