changes to training/monitor communication
the monitor now binds while the trainer connects, since we may have
a single monitor used for multiple trainers (in sequence). we use
the what field in the msg to ensure that the message came from the
trainer that we are currently running. the sleep in the event loop
is now shorter, since we were getting behind on messages.