Fix to Tracking when optical flow fails.
When optical flow fails shifted instances for tracks
are lost, causing no tracks to be found on previous
frames. This has been fixed so that if optical flow
is not working then we use the last know position
for the track.

There is still a bug that causes a new track to spawn
when a track doesn't exist on the previous frame cause
we did not match an instance. In this case I guess we
should go back for of_window number of frames looking
for an occurrence of the track. Or just keep a running
calculation per track for last known instance.