Disable contextual menu when clicking node.
Right-clicking on a node (or its label) toggles visibility, so
we don't want to show contextual menu (for adding instance).
Maybe we shouldn't use right-clicks for these two functions, but
it's not clear what a better UI for toggling node visibility would
be since this is something we want to be easy when labeling (i.e.,
we don't want to slow user down, and we already are using lots of
click+modifier for other things). The problem case is when user
wants to add a new instance that's on top of an existing node, but
this is probably rare and use can always just add an instance and
then move it where desired.