Add hog feature suggestion, better bag of features
Bag of features (with count vectors) used for brisk and hog
features. We generate the brisk keypoints, get all brisk/hog
descriptors for those keypoints, then cluster to get "vocab".
Each image is then given a vector with counts for each type of
keypoint descriptor (i.e. "vocab word") found in that image,
and these are then used for clustering the images.