Ensure that table sort doesn't break row idxs.
If you sort a data table, methods that rely on the correspondence
between row index and *original* items (i.e., not the dictionaries
we sometimes use to cache data from the original items) were
getting out of sync. For example, double-clicking a row in the
suggestions table would take you to the suggestion that had been
on that row before you sorted. This is now fixed and has test
coverage.

Fixes issue #278.