Always show horiz scroll in slider.
Now that it's possible to zoom, the horizontal scroll is useful
for showing whether you're zoomed in even when it's not needed
for scrolling. Since it's always shown we can also pad the scene
so that the appearance of the scrollbar doesn't cause the height
to no longer fit (which makes us then need a vertical scrollbar).