Improved background frame loading.
If the user held the mouse down (i.e., dragging over the seekbar),
then the request events weren't getting sent to the worker in a
timely way... so the were getting sent in chunks once the user
released the mouse, which meant the video frame could take a long
time to catch up once you actually stopped on a frame. Now the
requests are being sent directly through a method call rather than
an event, so there's no significant delay in queueing them up.

The worker also now uses a mutex lock to ensure that there's only
one frame (the most recent in the queue) being read at a time.
This should also help ensure we don't get a backlog of old frames
to show when the user stops changing frames.

These changes should resolve issue #295.