Bug fix when scrolling so slider handle in view.
If the slider area had been resized so that not all of the tracks
were in view (and the top or bottom of the handle was also not in
view), then moving the handle would make the view toggle between
scrolling to the top and scrolling to the bottom (since it was
trying to keep the handle in view and couldn't). Now we just
make sure the handle is in horizontal view and don't adjust the
vertical.