New dialog for complex instance deleting.
The dialog allows for deleting user/predicted/all instances
in specified frames/videos/tracks.

The old "delete from clip" command has been modified to delete all
predicted instances in the clip, regardless of whether an instance
is selected. Previously it would only delete instances which
matched the track of the selected instance if there was one.
This was confusiong, so the current change helps address
issue #312.