Use RLock around ffmpeg seek/read.
Without a lock we were getting errors when using viz during
training. With a non-recursive lock I get a deadlock when starting
inference. It appears that everything works with a recursive lock.

I don't see any noticable performance cost but we should keep an
eye out for this.

Fixes issue #321.