Use .slp as default for hdf5 datasets.
Users will still be able to save/open .h5 files but when saving
the default ext is now .slp (the actual format is unchanged).
This may require users to change some scripts that expected (eg)
predictions.h5 as the output filename from inference.

This is the first stage of issue #304.