Improvements to training/inference config export.
It should now be possible to directly run the training script
and then the inference script, with the parent directory set as
the current working dir.

All paths to external files (videos, labels, trained models) are
now absolute (so they may need to be fixed if running scripts on
a system with network drives mapped differently).

All paths to internal files (training configs, currently untrained
models, output predictions) are relative to the directory in which
config files are being exported. This allows user to run scripts
with this dir as the cwd.

Resolves issue #334.