Add (disabled) code for NMS during track cleaning.
If enabled by setting iou_threshold, then during track cleaning
we'll determine which instances over target count to remove by
considering instances which overlap (iou over threshold). We won't
remove *all* instances over threshold, only enough to bring the
number of instances in the frame down to the target count.

If there are still more instances in the frame than the target
number (or if iou_threshold is set to None), then we'll still
remove the lowest scoring instances to get down to the target
count (as before).