Accept list of new video paths in training cli.
sleap-train now has a --video-paths arg which should be a comma-
delimited list of video paths.

If Labels.load_file is given a list of video paths (via video_
search arg) and this list is the same length as the list of videos
and the file extensions match, then these new paths will be used
to replace paths of videos which can't be found (even if the
*filename* of the video is different).

All of this is to allow specifying a list of video paths to the
command line so that we can find videos if the paths have been
fixed in the loaded labels object but not in the saved labels file.

Fixes issue #335.